Control code major changes
Modified startrecipe.php to ask server to start a production. Server now checks the chosen recipe integrity and, if it is well enough, then the recipe production starts (not implemented yet). Also, there is a log file to save the recipe progress and a lockfile to tell any interested parts of the system that there is a recipe in progress or not